+++ cli/bin/aofl-cli.js
@@ -19,7 +19,8 @@ program
 .command('conclude <moduleName>', 'conclude module')
 .command('g [type] [name]', 'scoffold component')
 .command('sw [path]', 'generate service worker')
-.command('i18n <moduleName>', 'Generates .pot files from source');
+.command('i18n <moduleName>', 'Generates .pot files from source')
+.command('dom-scope [path]', 'validate generated dom-scope');
 
 program
 .parse(process.argv);