+++ component-utils/index.js
@@ -1,5 +1,3 @@
-import findParent from './src/find-parent';
-
-export {
-  findParent
-};
+export {default as findParent} from './src/find-parent';
+export {default as isInViewport} from './src/is-in-viewport';
+export {default as isInViewportMixin} from './src/is-in-viewport-mixin';