+++ web-components/aofl-img/documentation/routes/home-element/examples/scroll/template.js
@@ -1,0 +1,6 @@
+(context, html) => html`
+<div class="scroll-container">
+  <span class="placeholder"></span>
+  <aofl-img src="https://via.placeholder.com/500x500" width="500" height="500" alt="scroll to load img"></aofl-img>
+</div>
+`;