+++ web-components/aofl-img/documentation/routes/home-element/examples/visible/index.js
@@ -1,0 +1,3 @@
+export {default as jsCode} from '!raw-loader!./example';
+export {default as htmlCode} from '!raw-loader!./template';
+export {default as cssCode} from '!raw-loader!./index.css';