+++ web-components/aofl-img/documentation/routes/home-element/examples/visible/template.js
@@ -1,0 +1,3 @@
+(context, html) => html`
+<aofl-img src="https://via.placeholder.com/200x100" width="200" height="100" alt="demo img in visible area on load"></aofl-img>
+`;