+++ web-components/aofl-picture/documentation/routes/home-element/examples/picture/index.js
@@ -1,0 +1,3 @@
+export {default as jsCode} from '!raw-loader!./exmaple';
+export {default as cssCode} from '!raw-loader!./index.css';
+export {default as htmlCode} from '!raw-loader!./template';