+++ web-components/aofl-picture/documentation/webpack.dev.testing.js
@@ -1,0 +1,22 @@
+const merge = require('webpack-merge');
+const devConfig = require('./webpack.dev');
+const UnitTesting = require('@aofl/unit-testing-plugin');
+
+const config = merge(devConfig, {
+  plugins: [
+    new UnitTesting({
+      exclude: [
+        '**/node_modules',
+        '**/node_modules_sourced',
+        '**/js/__config'
+      ],
+      scripts: [
+        'runtime',
+        'init-polyfill-service',
+        'custom-elements-es5-adapter'
+      ]
+    })
+  ]
+});
+
+module.exports = config;