+++ web-components/aofl-source/template.js
@@ -1,3 +1,0 @@
-import {html} from '@polymer/lit-element';
-
-export const template = (context) => html``;